take filenames with dots into account in get_sigmf_filenames()
The function get_sigmf_filenames() does not work correctly if a
file contains dots in the filename. These are interpretted as suffixes
by pathlib, and a part of the filename can be removed when forming
the different SigMF filenames. This fixes the problem by only
considering as file extensions the canonical SigMF file extensions.
Any other suffixes are treated as part of the filename and not removed.

This fixes #106.