Take filenames with dots into account in get_sigmf_filenames() (#107)
* take filenames with dots into account in get_sigmf_filenames()

The function get_sigmf_filenames() does not work correctly if a
file contains dots in the filename. These are interpretted as suffixes
by pathlib, and a part of the filename can be removed when forming
the different SigMF filenames. This fixes the problem by only
considering as file extensions the canonical SigMF file extensions.
Any other suffixes are treated as part of the filename and not removed.

This fixes #106.

* fix SigMFCollection.get_SigMFFile()

The get_SigMFFile() method was broken because it used ".sigmf_meta"
instead of ".sigmf-meta". This rewrites the method to use
get_sigmf_filenames().

* test that SigMF files with . in the filename are loaded correctly

This adds a unit test that shows that SigMF files like
a.sigmf-meta, b.c.sigmf-meta, and d.e.f.sigmf-meta are all loaded
correctly.

* increment patch version

---------

Co-authored-by: Teque5 <teque5@gmail.com>